Carry user operations into new UsersController
