+++ API Project/src/main/java/hello/GreetingController.java
@@ -10,20 +10,6 @@ import java.util.ArrayList;
 public class GreetingController {
   private static final String template = "Hello, %s!";
   private final AtomicLong counter = new AtomicLong();
-  private ArrayList<User> users = new ArrayList<>();
-
-  @RequestMapping("/get-users")
-  public ArrayList<User> getUsers() {
-    return this.users;
-  }
-
-  @RequestMapping("/add-user")
-  public User addUser(@RequestParam(value="firstName", defaultValue="") String firstName,
-                      @RequestParam(value="lastName", defaultValue="") String lastName) {
-    User user = new User(firstName, lastName);
-    this.users.add(user);
-    return user;
-  }
 
   @RequestMapping("/greeting")
   public Greeting greeting(@RequestParam(value="name", defaultValue="World") String name) {
